Reorganizing global importance
